% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_dica.R
\name{fim_dica}
\alias{fim_dica}
\title{fim_dica}
\usage{
fim_dica(output = NULL)
}
\arguments{
\item{output}{'html' ou 'latex'. Se não fornecido, o formato de saída é
detectado automaticamente}
}
\value{
código HTML ou LaTeX
}
\description{
fim_dica
}
