% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_resposta.R
\name{fim_resposta}
\alias{fim_resposta}
\title{fim_resposta}
\usage{
fim_resposta(output = NULL, metadata = rmarkdown::metadata)
}
\arguments{
\item{output}{'html' or 'latex'. If not given, detect automatically.}

\item{metadata}{yaml metadata for document}
}
\value{
HTML or LaTeX code
}
\description{
End answer block
}
