% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniciar.R
\name{iniciar}
\alias{iniciar}
\title{Copy css and js files to local dir}
\usage{
iniciar(dir = "html_files")
}
\arguments{
\item{dir}{Target dir}
}
\value{
(Invisibly) Logical vector indicating which operation succeeded for each of the files attempted
}
\description{
Copy css and js files to local dir
}
