% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicio_dica.R
\name{inicio_dica}
\alias{inicio_dica}
\title{inicio_dica}
\usage{
inicio_dica(texto_botao = "Dica", output = NULL)
}
\arguments{
\item{texto_botao}{Button text}

\item{output}{'html' or 'latex'. If not given, detect automatically.}
}
\value{
HTML or LaTeX code
}
\description{
Begin a hint block
}
