% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicio_resposta.R
\name{inicio_resposta}
\alias{inicio_resposta}
\title{inicio_resposta}
\usage{
inicio_resposta(texto_botao = "Resposta", ancora = NULL, output = NULL)
}
\arguments{
\item{texto_botao}{Texto do botão}

\item{ancora}{Id da âncora HTML com a solução; USADO APENAS PARA SAÍDA PDF,
para gerar link para solução no site}

\item{output}{'html' ou 'latex'. Se não fornecido, o formato de saída é
detectado automaticamente}
}
\value{
código HTML ou LaTeX
}
\description{
inicio_resposta
}
