% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipla.R
\name{multipla}
\alias{multipla}
\title{multipla}
\usage{
multipla(opts, output = NULL)
}
\arguments{
\item{opts}{Choices see \link[webexercises:longmcq]{webexercises::longmcq}}

\item{output}{'html' or 'latex'. If not given, detect automatically.}
}
\value{
HTML or LaTeX code
}
\description{
Generate multiple choice question (long)
}
