% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_dica.R
\name{fim_dica}
\alias{fim_dica}
\title{fim_dica}
\usage{
fim_dica(output = NULL)
}
\arguments{
\item{output}{'html' or 'latex'. If not given, detect automatically.}
}
\value{
HTML or LaTeX code
}
\description{
End hint block
}
