% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_resposta.R
\name{fim_resposta}
\alias{fim_resposta}
\title{fim_resposta}
\usage{
fim_resposta(output = NULL)
}
\arguments{
\item{output}{'html' ou 'latex'. Se não fornecido, o formato de saída é
detectado automaticamente}
}
\value{
código HTML ou LaTeX
}
\description{
fim_resposta
}
