% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicio_dica.R
\name{inicio_dica}
\alias{inicio_dica}
\title{inicio_dica}
\usage{
inicio_dica(texto_botao = "Dica", output = NULL)
}
\arguments{
\item{texto_botao}{Texto do botão}

\item{output}{'html' ou 'latex'. Se não fornecido, o formato de saída é
detectado automaticamente}
}
\value{
código HTML ou LaTeX
}
\description{
inicio_dica
}
