% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicio_resposta.R
\name{inicio_resposta}
\alias{inicio_resposta}
\title{inicio_resposta}
\usage{
inicio_resposta(
  texto_botao = "Resposta",
  output = NULL,
  metadata = rmarkdown::metadata
)
}
\arguments{
\item{texto_botao}{Button text}

\item{output}{'html' or 'latex'. If not given, detect automatically.}

\item{metadata}{yaml metadata for document}
}
\value{
HTML or LaTeX code
}
\description{
Begin answer block
}
