% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacuna.R
\name{lacuna}
\alias{lacuna}
\title{lacuna}
\usage{
lacuna(
  answer,
  width = min(40, max(nchar(answer))),
  num = NULL,
  ignore_case = FALSE,
  tol = NULL,
  ignore_ws = TRUE,
  regex = FALSE,
  output = NULL
)
}
\arguments{
\item{answer}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{width}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{num}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{ignore_case}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{tol}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{ignore_ws}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{regex}{see \link[webexercises:fitb]{webexercises::fitb}}

\item{output}{'html' or 'latex'. If not given, detect automatically.}
}
\value{
HTML or LaTeX code
}
\description{
Generate fill-in-the-blanks exercise
}
