% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipla.R
\name{multipla}
\alias{multipla}
\title{Múltipla escolha}
\usage{
multipla(opts, output = NULL)
}
\arguments{
\item{opts}{Opções see \link[webexercises:longmcq]{webexercises::longmcq}}

\item{output}{'html' ou 'latex'. Se não fornecido, o formato de saída é
detectado automaticamente}
}
\value{
código HTML ou LaTeX
}
\description{
Múltipla escolha
}
